Use Illuminant in texture chaining
Reflectant will clamp Spectral to be 1, to avoid amplification in BSDF.
Illuminant will not.

When chaining multiple textures to compute a reflectant, you don't want
to use Reflectant for the intermediate Spectrals, because it alters the
results. It doesn't matter that the intermediates exceed 1, because only
the final Spectral must be clamped.