Add FilterTriangle to output
Though PBRT puts the default width of the triangle filter at 2,
I believe it should really be 1. Remember width is here half the extent,
so width is to extent like radius is to diameter. The width should be 1
so that when a sample hits a pixel position dead center, the filter
kernel should drop to zero for the neighbouring pixels. So in effect, we
bilinear filtering with a tent kernel.

Similarly, the width of the "box filter" is 0.5, so that it becomes a
nearest neighbour filter (this is also the default width for the box
filter in PBRT)