Import liar in test_driver
As global test environment, import liar into the test_driver process.

As tests run after the build (before Conan packages liar into a neat
importable package), we need some way to import liar from the build
artifacts.  The extension modules will be somewhere in the build tree,
and while we could copy the pure python modules and the data into the
build tree as well, to reflect the final liar package, it's nicer to
load them directly from the source tree. But that means we need to
virtually patch together all files into one package.

In comes `sys.meta_path` on which we can add a MetaPathFinder in which
we can handle all liar imports and redirect them to the right files.

We only need to handle the liar and liar.codecs packages themselves
(i.e. loading their `__init__.py` modules), because by setting
`submodule_search_locations` we can tell where the submodules are found
and which can be multiple directories: the src directory containing the
tools submodule, and the bin directory containing the extension modules.

To make sure `pkgutil.iter_modules(__path__)` in `liar.codecs` properly
works, we've moved the codecs extension modules in a subdirectory of
the bin dir, so that they can be properly listed by themselves.

Loading data needs special care too. This doesn't go through
LiarPathFinder itself, but is handled by the module's loader. Therefore,
we overloaded `get_data` in `LiarLoader` for the main liar package.

Need to Wrap std::codecvt as its desctructor is protected [1,2].
Without the wrapper, we get an error like this:

  error: ‘virtual std::codecvt<wchar_t, char, __mbstate_t>::~codecvt()’ is protected within this context

[1] https://en.cppreference.com/w/cpp/locale/codecvt/%7Ecodecvt
[2] https://en.cppreference.com/w/cpp/locale/codecvt