Fix IccSpace
This will be using the LCMS2 color engine to fully handle ICC profiles.

No need for CMS_DLL as we're linking to static library

We can't use cmsCreateXYZProfile to create the XYZ color space, as this
function creates one with a D50 Illuminant, while we need one with
E Illumant.  So we need to create an "RGB" colour space with proper
illumant, primaries _and_ tonecurves. Using nullptr for primaries and
tonecurves---like cmsCreateXYZProfile does---does not work:
cmsCreateTransform will fail to link the two colour spaces.