Extract RGB space from ICC space
You cannot directly use the colorants to derive the RGB primaries
(converting XYZ coordinates to xy): they have the wrong whitepoint.
Colorants in ICC profiles are refered to D50 illuminant, while you first
need to convert them to the colour space's original whitepoint
(e.g. D65, so explicitly not to E of the CIE XYZ we use internally).

The colorants in the ICC profile are basically the same as the r_, g_
and b_ XYZ colors we store in an RgbSpace, but than refered to D50
instead of E. So we need to undo the M_CAT matrix, then we can extract
the XYZ primaries, and finally map them on the xy plane.