Add RemapPbrtRoughness
It remaps PBRT's interpretation of roughness to LiAR's, ensuring the
same alpha values in the microfacet distributions will be used.

For PBRT, alpha=roughness. For LiAR, alpha=roughness^2, so we at least
need to take the square root. But if remaproughness is True, then PBRT
applies an additional RoughnessToAlpha function.