Decrease refreshing rate of Display
By removing the semaphore from RenderEngine::writeRender, it became
evident that we only had about 80% CPU usage. Before it seemed
100% because of the spinlock ...  The tonemapping of the Display refresh
is the culprit here, especially with the Halton sampler that basically
needs to refresh the whole screen every cycle.  By reducing the refresh
rate, we can increase CPU usage to well in 90%.