Use light strategy for glossy surfaces in AdjointPhotonTracer
It's only specular surfaces that don't benefit from sampling the lights
instead of only sampling the BSDF. Especially rough glossy surfaces
can really benefit from the light sampling.

This was made most obvious by the most right leather chair in the
barcelona-pavilion examle of pbrt-v3-scenes. The Fourier BSDF is
classified as glossy only, but this one is in fact quite diffuse.

Ideally, each BSDF would be able to return a single value between 0 and
1, telling how diffuse or specular it is.