Adjust differential ray for supersampling
When using supersampling (i.e. more than one sample per pixel), the
footprint of the differential ray can be reduced accordingly, as the
multiple samples taken will compensate.

This results in more crisp texture evaluation, and this is especially
important for bumpmapping, as noticed in the barcelona pavilion scene of
pbrt-v3-scenes. Without this reduction of the footprint, the
displacement map is evaluated and a too deep mipmap level for distant
pixels, so that the bumps become too flat.