Fix FilterTriangle
The kernel was completely wrong.

We also believe the default width should be 1, in contrary to PBRT. The
width is really the radius, not the extent of the filter. We believe
that by default, the weight should be 1 for the center, and drop to
zero at exactly one pixel away. So if a sample lands exactly on a pixel
position, it's fully contributed to that pixel and nothing is added to
the neighbours. But if a sample lands exactly between two pixels, each
pixel will get half the contribution. Also, for width=1, the area under
the pixel is exactly 1.

Compare this to the box filter that does have a default width 0.5 in
PBRT: the sample is fully contributed to the nearest pixel, and nothing is
added to the neighbouring ones, and the area is also 1 (unless the sample
lands exactly in between two pixels, then I suppose both get the full
contribution).

Also, the only occurrences of the triangle filter in the PBRT v3 scenes
always specifically set the width to 1.