Fix MIS in direct lighting for specular materials
When seeing an area light reflected in a specular surface like a mirror, we
in fact see the direct lighting contribution of that area light on this
surface (i.e. we didn't reflect the camera ray and traced it to the
area light where we added the surface emission; it's a shadow ray being
traced to the area light to add a light contribution).

Say nl=1 and nb=1 (nb!=0 because light is not singular).

Sampling the light (nl) will not add any contribution, as the surface is
specular and thus out.value will be black and out.pdf will be zero.

Sampling the BSDF (nb) will have some out.value > 0 and out.pdf > 0, and
also lightPdf could be > 0; in which case the contribution will be
weighted down by the power heuristic on nb * out.pdf and nl * out.pdf
which will be less than 1.

So the only sample that can add a contribution (the sampled BSDF) get
its contribution weighted down as we assume sampling the light could
also add valid contribution, which it can't as the BSDF is specular.

Conclusion: if the BSDF sample was specular, then it should not weight
it with the light PDF, and assume only the BSDF has been sampled.