Fix Observer CDF
CDF had one element more that wavelengths, that could lead to reading w_
beyond it's end.

The CDF(w) is the integral of pdf(w) from w[0] to w. We choose pdf(w)
to be flat between w[i] and w[i+1], so that CDF(w) is piecewise-linear,
and we only need to know it in w[i]:

CDF[i] = CDF[i-1] + pdf[i-0.5] * (w[i] - w[i-1])
CDF[0] = 0

pdf[i-0.5] is pdf(w) between w[i-1] and w[i], and is choosen to be
proportional to abs total of the average xyz(w) in that interval.