Fix infinite dPoint_dV in LightSky
If dir.z == 1 (dir.x and dir.y can be zero but don't have to be),
then cosTheta == 1 and sinTheta == 0, and cosTheta_sinTheta = inf.
This causes dNormal_dV and dPoint_dV to be infinite, eventually causing
NaNs in IntersectionContext::bsdfToLocal(), blowing everything up:

  TVector3D v = normal_.reject(dPoint_dV_)

If dir.z == 1, then dir.x and dir.y _should_ be zero, and dDir_dU and
dDir_dV are in fact undefined, so anything goes. Therefore, we'll set it
to (0, 0, 0), and will assume downstream code will handle this ...