Explicitly store worldToLocal in Transformation
Profiling showed quite some time being spent in
Transformation3D::inverse() being called from
TTransformation3D::doIntersect. Not in the top 5 of hot functions, but
maybe in top 10.  This is odd, because the inverse is only calculated
once and then cached (and shared over all threads).

It seems the atomic read to check if the inverse is already present might
be the problem here.  By explictly storing the inverse transformation,
we can avoid the repeated inverse() call, and thus the atomic read,
getting us from about 19 min to 18 min on the barcelona pavillion.