Fix crash in texture lookup due to NaN in direction
In the image texture lookup of the skymap, an crash could be observed
because an index went out of range (-1 < 0), due to one of the
coordinates being NaN. This originated from the z-coordinate of the
direction being NaN, which originated from a 0 / 0 division. This can
happen if sample.x or sample.y equals zero.

The crash was observed with the pbrt v3 barcelona-pavilion scene