Align Image pixel blocks on cache-lines.
Make sure that the blocks used to organise the pixels are cache-aligned.
This improves lookup speed. On the Barcelona Pavillion test scene, this
goes from 8m01s to 7m55s with clang compiler and virtualization
disabled. (the jump from the 18m runtimes before are due to running the
tests on a newer CPU)