Use faster log2 in Image::lookUp
In Image::lookUp, we use log2 to select the right mip-map level, and in
case of trilinear interopolation, we use the fractional part to blend
between two levels.

In case of bilinear interpolation or no interpolation (nearest
neighbour), we don't even need the fractional part and only need the
integer part of the result, which is exactly the exponent of IEEE 754
base-2 floating point representation, which we can easily extract from
the binary representation:

  w = 2^e * m

In case of trilinear interpolation, the blend factor dLevel is in fact
log2(m), which we can apprimate by (m-1) with a maximum error of 0.09.
This m can also easily be extracted from the binary representation of w.

For more information, see notes/mipmap_log2_approx.ipynb

On the PBRT v3 Barcelona pavilion test scene, this reduced the runtime
from 7m55s to 7m51s.