Fix Inf in Sphere::doLocalContext with normal.z==1
Fixes following assertion failure:

  ASSERT FAILURE: C:\coding\liar\src\kernel\intersection_context.h(155): liar::kernel::IntersectionContext::setDNormal_dV: ::liar::kernel::isFinite(dNormal_dV): (inf, inf, -0)

Local context was being computed for a normal with z=1 exactly.
This caused sinTheta to be 0, which caused cosTheta_sinTheta to be Inf.
This eventually resulted in dNormal_dV having Inf values in x and y.
These could also have been NaNs, but normal.x and normal.y were not 0 in
this case.