Remove liar.tools.colors
The way it tries to create a wrapper on RgbSpace.setDefaultSpace can't
work as you can't assign a method to a static type.

We could create a dedicated hook mechanism to add hooks that are called
from the C++ side when a new default RGB space is set, but this module
isn't being used anywhere anyway, so let's just remove it.