Prevent using -imsvc for Clang to fix IntelliSense
Unset CMAKE_INCLUDE_SYSTEM_FLAG_CXX to prevent CMake from using the
-imsvc option instead of /I for third-party include headers (of imported
targets) when generating VS solution files using the Clang-CL compiler,
to workaround a bug in VS IntelliSense that would not resolve symbols
or headers if the -imsvc option was being used.

Instead of the normal -I option, GCC and Clang have the -isystem [1]
option for include directories of system libraries. One of the effects
is that for headers found in these system include directories, all
warnings other than #warning are suppressed [2].

Visual Studio gained a similar option /external:I in VS2017 v15.6 [3],
which was later adopted by CMake to be used for all third-party headers
(i.e. for imported targets of external dependencies) [4].

This works great when using the MSVC compiler, but when using Clang-CL
this breaks IntelliSense (I'm using VS Community 2022 17.13.6), much
like this bug report [5] that claims this to have been fixed a long time
ago). The code build just fine, but using F12 simply states that headers
or symbols can't be found, which results in a terrible developer
experience.

When using Clang-CL, CMake uses the -imsvc option instead of /external:I
when generating the VS solution files (as tested with CMake 3.26.5 and
3.31.6). Manually changing it to /external:I doesn't resolve the issue
though, so it seems to go deeper than that.

One workaround that I've found is to simply disable the usage of any
special option for third-party headers by unsetting the apparently
undocumented CMAKE_INCLUDE_SYSTEM_FLAG_CXX [6], which is exactly what
I'm doing right now.

PS: VS Code does not suffer the same fate, because we already switch
IntelliSense in windows-clang-x64 mode in .vscode/c_cpp_properties.json.
So it may not be IntelliSense per-se, but Visual Studio IDE that doesn't
switch the mode when loading Clang-CL projects. But then again,
/external:I doesn't work either with Clang-CL projects, so 🤷 ...

[1] https://gcc.gnu.org/onlinedocs/gcc/Directory-Options.html
[2] https://gcc.gnu.org/onlinedocs/cpp/System-Headers.html
[3] https://devblogs.microsoft.com/cppblog/broken-warnings-theory/
[4] https://gitlab.kitware.com/cmake/cmake/-/issues/17904
[5] https://developercommunity.visualstudio.com/t/Go-To-Document-F12-for-C-include/1191078
[6] https://discourse.cmake.org/t/imported-projects-dont-resolve-include-paths-in-intellisense/9185