Use a not-nullable PyObjectRef (prototype)
A PyObjectPtr<T>::Type is a smart pointer to T that can hold a null
pointer, or be None—the equivalent in Python. But this is often not what
we want: most PyObjectPtr pointers in liar are assumed to always hold an
instance.

Now that we are generating *.pyi stub files, this is obvious by the many
"T | None" unions in method parameters or return types. For example:

  class RgbSpace:
      ...
      @classsmethod
      def defaultSpace(cls, /) -> RgbSpace | None:
          ...
      @classmethod
      def setDefaultSpace(cls, defaultSpace: RgbSpace | None, /) -> None:
          ...

This also results in mypy errors such as:

  examples\raytracinginaweekend\raytracinginaweekend.py:19: error: Item "None" of "RgbSpace | None" has no attribute "linearSpace" [union-attr]
  examples\raytracinginaweekend\raytracinginaweekend.py:85: error: Item "None" of "RgbSpace | None" has no attribute "withGamma"  [union-attr]

RgbSpace *must* have a default space, it doesn't make sense for it to be
None. In fact, liar will crash if it is None, because the code doesn't
check for null pointers before dereferencing it. And like this, there
are many, many more examples.

To mitigate this—both the T | None typing and the lack of null-pointer
checking before dereferencing—I've prototyped a wrapper around
PyObjectPtr called PyObjectRef<T>, short for PyObject reference.
Much like C++ references, they can't be null (*), but unlike C++
references, they can be reassigned. This prototype is inspired by
gsl::not_null [1, 2]

This allows us to specialize PyExportTraits for PyObjectRef<T> to
explicitly check if None is being passed in, and raise a TypeError
before this is converted to a null pointer. It also allows us to simply
have "T" as Python type hint, instead of "T | None", so that this
correctly shows up in the *.pyi stubs.

We obviously need to be able to convert PyObjectPtr<T>::Type to
PyObjectRef, and those can be null. We cannot prevent this at
compile-time, but we can enforce that they are not null at runtime, and
throw an exception if they are.

(*) There's one problem: to be able to use PyObjectRef<T> in containers
like std::vector, we need to be able to default-construct them. Not for
the containers per-se, but for the way these containers are exported to
Python and how they get their elements from Python. The signature of
pyGetSimpleObject takes the element as reference:

  template <typename T> int pyGetSimpleObject(PyObject* iV, T& oV)

and elements are default contructed first:

  typename Container::value_type value;
  if (pyGetSimpleObject(obj,value) != 0) { ... }

So we have to allow PyObjectRef<T> to be default-constructed containing
a null pointer, puncturing a hole in our defense. But with some
additional runtime checks on copy construction and friends, we can at
least mitigate the worst problems with little additional cost—and it's
not like we were checking for null pointers before anyway.

For a better solution that avoids needing default construction, we'll
need some major changes in Lass that can directly return the value
from pyGetSimpleObject, instead of passing it by reference.

[1] https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Ri-nullptr
[2] https://visualstudiomagazine.com/articles/2016/06/01/using-the-not_null-template.aspx